refactor migration diff helper boundaries
