+++ .sampo/changesets/issue-367-migration-diff-helpers.md
@@ -1,0 +1,5 @@
+---
+npm/@wp-typia/project-tools: patch
+---
+
+Split migration diff rename and transform helpers into focused runtime modules without changing diff output.
